// ai-notes-backend/src/enums/notification-type.enum.ts
export enum NotificationType {
  NOTE_SHARED = 'NOTE_SHARED',
  NOTE_ACCESS_REVOKED = 'NOTE_ACCESS_REVOKED',
  SHARE_LINK_CLAIMED = 'SHARE_LINK_CLAIMED',
  NOTE_INVITE = 'NOTE_INVITE',
  PROJECT_SHARED = 'PROJECT_SHARED',
  PROJECT_ACCESS_REVOKED = 'PROJECT_ACCESS_REVOKED',
  PROJECT_SHARE_LINK_CLAIMED = 'PROJECT_SHARE_LINK_CLAIMED',
  PROJECT_INVITE = 'PROJECT_INVITE',
}